+++ sdk/python/lib/pulumi/stack_reference.py
@@ -11,7 +11,8 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from typing import Optional, Any
+from asyncio import ensure_future
+from typing import Optional, Any, List
 from copy import deepcopy
 
 from .output import Output, Input
@@ -34,6 +35,11 @@ class StackReference(CustomResource):
     The outputs of the referenced stack.
     """
 
+    secret_outputs: Output[List[str]]
+    """
+    The names of any stack outputs which contain secrets.
+    """
+
     def __init__(self,
                  name: str,
                  stack_name: Optional[str] = None,
@@ -52,6 +58,7 @@ class StackReference(CustomResource):
         super().__init__("pulumi:pulumi:StackReference", name, {
             "name": target_stack,
             "outputs": None,
+            "secret_outputs": None,
         }, opts)
 
     def get_output(self, name: Input[str]) -> Output[Any]:
@@ -60,7 +67,10 @@ class StackReference(CustomResource):
 
         :param Input[str] name: The name of the stack output to fetch.
         """
-        return Output.all(Output.from_input(name), self.outputs).apply(lambda l: l[1].get(l[0]))
+        value = Output.all(Output.from_input(name), self.outputs).apply(lambda l: l[1].get(l[0]))
+        is_secret = ensure_future(self.__is_secret_name(name))
+
+        return Output(value.resources(), value.future(), value.is_known(), is_secret)
 
     def require_output(self, name: Input[str]) -> Output[Any]:
         """
@@ -69,4 +79,36 @@ class StackReference(CustomResource):
 
         :param Input[str] name: The name of the stack output to fetch.
         """
-        return Output.all(Output.from_input(name), self.outputs).apply(lambda l: l[1][l[0]])
+
+        value = Output.all(Output.from_input(name), self.outputs).apply(lambda l: l[1][l[0]])
+        is_secret = ensure_future(self.__is_secret_name(name))
+
+        return Output(value.resources(), value.future(), value.is_known(), is_secret)
+
+    def translate_output_property(self, prop: str) -> str:
+        """
+        Provides subclasses of Resource an opportunity to translate names of output properties
+        into a format of their choosing before writing those properties to the resource object.
+
+        :param str prop: A property name.
+        :return: A potentially transformed property name.
+        :rtype: str
+        """
+
+        return "secret_outputs" if prop == "secretOutputs" else prop
+
+    async def __is_secret_name(self, name: Input[str]) -> bool:
+        # If either the name or set of secret outputs is unknown, we can't do anything smart, so we
+        # just copy the secretness from the entire outputs value.
+        if not (await Output.from_input(name).is_known() and await self.secret_outputs.is_known()):
+            return await self.outputs.is_secret()
+
+        # Otherwise, if we have a list of outputs we know are secret, we can use that list to
+        # determine if this output should be secret. Names could be None here in cases where we are
+        # using an older CLI that did not return this information (in this case we again fallback to
+        # the secretness of outputs value).
+        names = await (self.secret_outputs.future())
+        if names is None:
+            return await self.outputs.is_secret()
+
+        return await Output.from_input(name).future() in names