+++ tests/integration/stack_reference_secrets/step1/index.ts
@@ -1,0 +1,5 @@
+import * as pulumi from "@pulumi/pulumi";
+
+export const normal = pulumi.output("normal");
+export const secret = pulumi.secret("secret");
+